Move harbors types into types.ts
