+++ src/domain/harbors/utils.ts
@@ -1,29 +1,5 @@
 import { HARBORS } from './__generated__/HARBORS';
-
-type Map = {
-  id: string;
-  url: string;
-};
-
-export interface HarborData {
-  id: string;
-  electricity: number;
-  gate: number;
-  lighting: number;
-  wasteCollection: number;
-  water: number;
-  name: string;
-  numberOfPlaces: number;
-  numberOfFreePlaces: number;
-  streetAddress: string | null;
-  zipCode: string | null;
-  municipality: string | null;
-  wwwUrl: string | null;
-  maps: Map[];
-  imageFile: string | null;
-  servicemapId: string | null;
-  maxWidth: number | null;
-}
+import { HarborData, Map } from './types';
 
 export const getHarborsData = (data: HARBORS | undefined) => {
   if (data?.harbors?.edges) {
